+++ src/fidesops/ops/schemas/saas/saas_config.py
@@ -264,6 +264,7 @@ class SaaSType(Enum):
     datadog = "datadog"
     rollbar = "rollbar"
     braze = "braze"
+    firebase_auth = "firebase_auth"
 
 
 class SaaSConfigBase(BaseModel):