+++ src/fidesops/ops/service/saas_request/__init__.py
@@ -1,3 +1,4 @@
 from fidesops.ops.service.saas_request.override_implementations import (
+    firebase_auth_request_overrides,
     mailchimp_request_overrides,
 )