+++ tests/ops/api/v1/endpoints/test_connection_template_endpoints.py
@@ -48,7 +48,7 @@ class TestGetConnections:
         resp = api_client.get(url, headers=auth_header)
         data = resp.json()["items"]
         assert resp.status_code == 200
-        assert len(data) == 25
+        assert len(data) == 26
 
         assert {
             "identifier": ConnectionType.postgres.value,
@@ -82,7 +82,7 @@ class TestGetConnections:
         resp = api_client.get(url + "?search=re", headers=auth_header)
         assert resp.status_code == 200
         data = resp.json()["items"]
-        assert len(data) == 3
+        assert len(data) == 4
         assert data == [
             {
                 "identifier": ConnectionType.postgres.value,
@@ -93,6 +93,11 @@ class TestGetConnections:
                 "identifier": ConnectionType.redshift.value,
                 "type": SystemType.database.value,
                 "human_readable": "Amazon Redshift",
+            },
+            {
+                "identifier": SaaSType.firebase_auth.value,
+                "type": SystemType.saas.value,
+                "human_readable": "Firebase Auth",
             },
             {
                 "identifier": SaaSType.outreach.value,
@@ -124,7 +129,7 @@ class TestGetConnections:
         resp = api_client.get(url + "?search=Re", headers=auth_header)
         assert resp.status_code == 200
         data = resp.json()["items"]
-        assert len(data) == 3
+        assert len(data) == 4
         assert data == [
             {
                 "identifier": ConnectionType.postgres.value,
@@ -135,6 +140,11 @@ class TestGetConnections:
                 "identifier": ConnectionType.redshift.value,
                 "type": SystemType.database.value,
                 "human_readable": "Amazon Redshift",
+            },
+            {
+                "identifier": SaaSType.firebase_auth.value,
+                "type": SystemType.saas.value,
+                "human_readable": "Firebase Auth",
             },
             {
                 "identifier": SaaSType.outreach.value,
@@ -152,7 +162,7 @@ class TestGetConnections:
         resp = api_client.get(url + "?system_type=saas", headers=auth_header)
         assert resp.status_code == 200
         data = resp.json()["items"]
-        assert len(data) == 15
+        assert len(data) == 16
 
         resp = api_client.get(url + "?system_type=database", headers=auth_header)
         assert resp.status_code == 200