+++ tests/ops/conftest.py
@@ -48,6 +48,7 @@ from .fixtures.saas.datadog_fixtures import *
 from .fixtures.saas.hubspot_fixtures import *
 from .fixtures.saas.mailchimp_fixtures import *
 from .fixtures.saas.outreach_fixtures import *
+from .fixtures.saas.request_override.firebase_auth_fixtures import *
 from .fixtures.saas.request_override.mailchimp_override_fixtures import *
 from .fixtures.saas.rollbar_fixtures import *
 from .fixtures.saas.salesforce_fixtures import *